requirements.txt
altair==4.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
chembl-structure-pipeline==1.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.29
greenlet==1.1.3
idna==3.4
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==5.0.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
lightgbm==3.3.2
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
munkres==1.1.4
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.3
plotly==5.10.0
prompt-toolkit==3.0.31
protobuf==3.20.3
pure-eval==0.2.2
pyaml==21.10.1
pyarrow==9.0.0
pydeck==0.8.0b4
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
reportlab==3.5.68
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.2
scikit-optimize==0.9.0
scipy==1.9.3
seaborn==0.12.0
semver==2.13.0
setuptools==65.5.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.42
stack-data==0.5.1
streamlit==1.13.0
streamlit-aggrid==0.3.3
streamlit-on-Hover-tabs==1.0.1
tenacity==8.1.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
unicodedata2==14.0.0
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
wheel==0.37.1
zipp==3.9.0
rdkit==2022.9.1
